requirements.txt
# iWatcher Partner - Python Dependencies

# API Clients
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
click>=8.1.0
colorama>=0.4.6
rich>=13.5.0

# Development
pytest>=7.4.0
black>=23.7.0